+++ flowcraft/generator/components/assembly_processing.py
@@ -204,6 +204,17 @@ class Pilon(Process):
         self.link_end.append({"link": "SIDE_BpCoverage",
                               "alias": "SIDE_BpCoverage"})
 
+        self.params = {
+            "clearInput": {
+                "default": "false",
+                "description":
+                    "Permanently removes temporary input files. This option "
+                    "is only useful to remove temporary files in large "
+                    "workflows and prevents nextflow's resume functionality. "
+                    "Use with caution."
+            }
+        }
+
         self.directives = {
             "pilon": {
                 "cpus": 4,