+++ flowcraft/generator/components/metagenomics.py
@@ -4,6 +4,7 @@ try:
 except ImportError:
     from flowcraft.generator.process import Process
 
+
 class Kraken(Process):
     """kraken process template interface
 
@@ -80,6 +81,14 @@ class MaxBin2(Process):
                 "default": 0.9,
                 "description": "probability threshold for EM final classification."
                                "Default: 0.9"
+            },
+            "clearInput": {
+                "default": "false",
+                "description":
+                    "Permanently removes temporary input files. This option "
+                    "is only useful to remove temporary files in large "
+                    "workflows and prevents nextflow's resume functionality. "
+                    "Use with caution."
             }
         }
 
@@ -129,6 +138,14 @@ class Megahit(Process):
                     "from the maximum read length of each assembly. If "
                     "'default', megahit will use the default k-mer lengths. "
                     "(default: $params.megahitKmers)"
+            },
+            "clearInput": {
+                "default": "false",
+                "description":
+                    "Permanently removes temporary input files. This option "
+                    "is only useful to remove temporary files in large "
+                    "workflows and prevents nextflow's resume functionality. "
+                    "Use with caution."
             }
         }
 
@@ -173,6 +190,14 @@ class Metaspades(Process):
                     "from the maximum read length of each assembly. If "
                     "'default', metaSPAdes will use the default k-mer lengths. "
                     "(default: $params.metaspadesKmers)"
+            },
+            "clearInput": {
+                "default": "false",
+                "description":
+                    "Permanently removes temporary input files. This option "
+                    "is only useful to remove temporary files in large "
+                    "workflows and prevents nextflow's resume functionality. "
+                    "Use with caution."
             }
         }
 
@@ -245,6 +270,14 @@ class RemoveHost(Process):
                 "default": "'/index_hg19/hg19'",
                 "description": "Specifies the reference indexes to be provided "
                                "to bowtie2."
+            },
+            "clearInput": {
+                "default": "false",
+                "description":
+                    "Permanently removes temporary input files. This option "
+                    "is only useful to remove temporary files in large "
+                    "workflows and prevents nextflow's resume functionality. "
+                    "Use with caution."
             }
         }